Testing and review fixes

+++ src/main/resources/db/migration/sqlserver/V2.7.1.20190405124500__split_output_files.sql
@@ -1,5 +1,5 @@
 CREATE TABLE ${ohdsiSchema}.output_file_contents (
-  output_file_id INTEGER,
+  output_file_id BIGINT,
   file_contents VARBINARY(max),
   CONSTRAINT output_file_contents_pkey PRIMARY KEY (output_file_id)
 )
